Treat RETRY as FAILURE
When testing kernel modules with buildbot it's entirely possible that
the system may panic resulting in a disconnect.  For the purposes of
automated testing we want to consider this a FAILURE and immediately
move on to testing the next change.

This change should not be pushed upstream until the behavior is made
more flexible.  For example, this could be done using a 'flunkOnRetry'
build step option.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>